Move some datafusion-optimizer::utils down to datafusion-expr::utils
These utils manipulate `LogicalPlan`s and `Expr`s and may be useful in
projects that only depend on `datafusion-expr`
